Third commit

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -9,9 +9,7 @@ import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class FilmService {
@@ -25,53 +23,47 @@ public class FilmService {
         this.userStorage = userStorage;
     }
 
-    public List<Film> getAllFilms() {
-        return filmStorage.getAllFilms();
-    }
-
     public Film createFilm(Film film) {
         return filmStorage.addFilm(film);
     }
 
     public Film updateFilm(Film film) throws FilmNotFoundException {
-        getFilmById(film.getId()); // Проверка наличия фильма
+        if (filmStorage.getFilmById(film.getId()) == null) {
+            throw new FilmNotFoundException("Фильм с ID " + film.getId() + " не найден");
+        }
         return filmStorage.updateFilm(film);
     }
 
-    public Film getFilmById(int id) throws FilmNotFoundException {
-        Film film = filmStorage.getFilmById(id);
+    public void deleteFilm(int filmId) throws FilmNotFoundException {
+        if (filmStorage.getFilmById(filmId) == null) {
+            throw new FilmNotFoundException("Фильм с ID " + filmId + " не найден");
+        }
+        filmStorage.deleteFilm(filmId);
+    }
+
+    public Film getFilmById(int filmId) throws FilmNotFoundException {
+        Film film = filmStorage.getFilmById(filmId);
         if (film == null) {
-            throw new FilmNotFoundException("Фильм с ID " + id + " не найден");
+            throw new FilmNotFoundException("Фильм с ID " + filmId + " не найден");
         }
         return film;
     }
 
+    public List<Film> getAllFilms() {
+        return filmStorage.getAllFilms();
+    }
+
     public void addLike(int filmId, int userId) throws FilmNotFoundException, UserNotFoundException {
-        if (userStorage.getUserById(userId) == null) {
-            throw new UserNotFoundException("User with id " + userId + " not found");
-        }
-        Film film = getFilmById(filmId);
-        film.addLike(userId);
-        filmStorage.updateFilm(film);
+        // Проверка существования фильма и пользователя делегирована хранилищу
+        filmStorage.addLike(filmId, userId);
     }
 
     public void deleteLike(int filmId, int userId) throws FilmNotFoundException, UserNotFoundException {
-        if (userStorage.getUserById(userId) == null) {
-            throw new UserNotFoundException("User with id " + userId + " not found");
-        }
-        Film film = getFilmById(filmId);
-        film.deleteLike(userId);
-        filmStorage.updateFilm(film);
+        // Проверка существования фильма и пользователя делегирована хранилищу
+        filmStorage.deleteLike(filmId, userId);
     }
 
     public List<Film> getPopularFilms(int count) {
-        return filmStorage.getAllFilms().stream()
-                .sorted(Comparator.comparingInt(Film::getLikesCount).reversed())
-                .limit(count)
-                .collect(Collectors.toList());
-    }
-
-    public void deleteFilm(int filmId) throws FilmNotFoundException {
-        filmStorage.deleteFilm(filmId);
+        return filmStorage.getPopularFilms(count);
     }
 }
